Extract queryJoin and queryExit functions to helper file
