+++ test/ComposableStable.integration.spec.ts
@@ -1,20 +1,16 @@
 // yarn test:only test/composableStable.integration.spec.ts
 import dotenv from 'dotenv';
 import { JsonRpcProvider } from '@ethersproject/providers';
-import { defaultAbiCoder } from '@ethersproject/abi';
-import {
-    Vault__factory,
-    BalancerHelpers__factory,
-} from '@balancer-labs/typechain';
+import { Vault__factory } from '@balancer-labs/typechain';
 import { vaultAddr } from './testScripts/constants';
 import { SubgraphPoolBase, SwapTypes, SOR, bnum } from '../src';
 import { Network, ADDRESSES } from './testScripts/constants';
 import { AddressZero } from '@ethersproject/constants';
-import { BigNumber, parseFixed } from '@ethersproject/bignumber';
+import { parseFixed } from '@ethersproject/bignumber';
 import { expect } from 'chai';
 import { closeTo } from './lib/testHelpers';
 import { ComposableStablePool } from '../src/pools/composableStable/composableStablePool';
-import { setUp } from './testScripts/utils';
+import { setUp, queryJoin, querySingleTokenExit } from './testScripts/utils';
 
 dotenv.config();
 
@@ -81,76 +77,6 @@ const testPool: SubgraphPoolBase = {
     ],
     amp: '1472',
 };
-
-export async function queryJoin(
-    network: number,
-    poolId: string,
-    assetsWithBpt: string[],
-    amountsInWithBpt: string[],
-    bptIndex: number
-): Promise<
-    [BigNumber, BigNumber[]] & { bptOut: BigNumber; amountsIn: BigNumber[] }
-> {
-    const helpers = BalancerHelpers__factory.connect(
-        ADDRESSES[network].balancerHelpers,
-        provider
-    );
-    const EXACT_TOKENS_IN_FOR_BPT_OUT = 1; // Alternative is: TOKEN_IN_FOR_EXACT_BPT_OUT
-    const minimumBPT = '0';
-    const abi = ['uint256', 'uint256[]', 'uint256'];
-    // ComposableStables must have no value for BPT in user data
-    const amountsWithOutBpt = [...amountsInWithBpt];
-    amountsWithOutBpt.splice(bptIndex, 1);
-    const data = [EXACT_TOKENS_IN_FOR_BPT_OUT, amountsWithOutBpt, minimumBPT];
-    const userDataEncoded = defaultAbiCoder.encode(abi, data);
-    const joinPoolRequest = {
-        assets: assetsWithBpt,
-        maxAmountsIn: amountsInWithBpt, // Must include BPT
-        userData: userDataEncoded,
-        fromInternalBalance: false,
-    };
-    const query = await helpers.queryJoin(
-        poolId,
-        AddressZero, // Not important for query
-        AddressZero,
-        joinPoolRequest
-    );
-    return query;
-}
-
-export async function querySingleTokenExit(
-    network: number,
-    poolId: string,
-    assetsWithBpt: string[],
-    bptAmountIn: string,
-    exitTokenIndex: number
-): Promise<
-    [BigNumber, BigNumber[]] & { bptIn: BigNumber; amountsOut: BigNumber[] }
-> {
-    const helpers = BalancerHelpers__factory.connect(
-        ADDRESSES[network].balancerHelpers,
-        provider
-    );
-    const EXACT_BPT_IN_FOR_ONE_TOKEN_OUT = 0; // Alternative is: BPT_IN_FOR_EXACT_TOKENS_OUT (No proportional)
-    const abi = ['uint256', 'uint256', 'uint256'];
-
-    const data = [EXACT_BPT_IN_FOR_ONE_TOKEN_OUT, bptAmountIn, exitTokenIndex];
-    const userDataEncoded = defaultAbiCoder.encode(abi, data);
-
-    const exitPoolRequest = {
-        assets: assetsWithBpt,
-        minAmountsOut: assetsWithBpt.map(() => '0'),
-        userData: userDataEncoded,
-        toInternalBalance: false,
-    };
-    const query = await helpers.queryExit(
-        poolId,
-        AddressZero, // Not important for query
-        AddressZero,
-        exitPoolRequest
-    );
-    return query;
-}
 
 describe('ComposableStable', () => {
     // Setup chain
@@ -323,11 +249,10 @@ describe('ComposableStable', () => {
                     pool._calcBptOutGivenExactTokensIn(amountsWithOutBpt);
 
                 const deltas = await queryJoin(
-                    networkId,
+                    provider,
                     testPool.id,
                     testPool.tokensList,
-                    amountsInWithBpt.map((a) => a.toString()),
-                    bptIndex
+                    amountsInWithBpt.map((a) => a.toString())
                 );
                 expect(bptCalculated.toString()).to.eq(
                     deltas.bptOut.toString()
@@ -353,11 +278,10 @@ describe('ComposableStable', () => {
                     pool._calcBptOutGivenExactTokensIn(amountsWithOutBpt);
 
                 const deltas = await queryJoin(
-                    networkId,
+                    provider,
                     testPool.id,
                     testPool.tokensList,
-                    amountsInWithBpt.map((a) => a.toString()),
-                    bptIndex
+                    amountsInWithBpt.map((a) => a.toString())
                 );
                 expect(bptCalculated.toString()).to.eq(
                     deltas.bptOut.toString()
@@ -390,7 +314,7 @@ describe('ComposableStable', () => {
                 );
 
                 const deltas = await querySingleTokenExit(
-                    networkId,
+                    provider,
                     testPool.id,
                     testPool.tokensList,
                     bptInEvm.toString(),
@@ -424,7 +348,7 @@ describe('ComposableStable', () => {
             //     );
 
             //     const deltas = await querySingleTokenExit(
-            //         networkId,
+            //         provider,
             //         pool.id,
             //         pool.tokensList,
             //         bptInEvm.toString(),
